requirements.txt
pandas
python-docx
fastapi
uvicorn
python-multipart

